TransactionsServices-main/central/requirements.txt
kafka-python
fastapi[standard]
psycopg[binary]